requirements.txt
torch
transformers
scikit-learn
datasets
jsonlines
OpenAttack
umap-learn
wget
sentence_transformers
strsimpy
hdbscan
language-tool-python
matplotlib
seaborn
peft
protobuf
bigmodelvis
torch-dct
